Add sounds and music
Use SDL_mixer when available.

We'll need to create a screen to give credit for the sounds and music.
The music is from
http://modarchive.org/index.php?request=view_by_moduleid&query=154578
(I've contacted Axel Ovekrans, the author, and he's granted permission
to use it), while for the sounds the information has been added to the
"Source" key of the metadata. All sounds are CC-licenced (most with
attribution clause, and all allow commercial usage).